test: add cmd test helper "newTestContext"
