+++ cmd/jaas/cmd/bootstrap_unit_test.go
@@ -3,13 +3,11 @@
 package cmd
 
 import (
-	"context"
 	"errors"
 	"fmt"
 	"testing"
 
 	qt "github.com/frankban/quicktest"
-	"github.com/juju/cmd/v3"
 	"github.com/juju/cmd/v3/cmdtesting"
 	"github.com/juju/gnuflag"
 	jujucloud "github.com/juju/juju/cloud"
@@ -137,7 +135,6 @@ func TestBootstrapRunDetached(t *testing.T) {
 	c.Assert(err, qt.IsNil)
 
 	f := gnuflag.NewFlagSet("test", gnuflag.ExitOnError)
-	f.SetOutput(s.writer)
 	command.SetFlags(f)
 	command.controllerName = "controller-name"
 	command.cloud = cloudName
@@ -146,11 +143,7 @@ func TestBootstrapRunDetached(t *testing.T) {
 	command.config = configOpts
 	command.detach = true
 
-	ctx := &cmd.Context{
-		Context: context.Background(),
-		Stdout:  s.writer,
-	}
-
+	ctx := newTestContext(t)
 	err = command.Run(ctx)
 	c.Assert(err, qt.IsNil)
 }
@@ -175,16 +168,6 @@ func TestBootstrapWatchLogs(t *testing.T) {
 		Watermark: 2,
 	}, nil)
 
-	s.writer.EXPECT().Write(gomock.Any()).DoAndReturn(func(b []byte) (int, error) {
-		c.Check(string(b), qt.Equals, "log-line\n")
-		return len(b), nil
-	}).Times(2)
-
-	s.writer.EXPECT().Write(gomock.Any()).DoAndReturn(func(b []byte) (int, error) {
-		c.Check(string(b), qt.Equals, "Job completed successfully.\n")
-		return len(b), nil
-	})
-
 	command := &bootstrapCommand{
 		store: s.store,
 		bootstrapAPIFunc: func() (JIMMAPI, error) {
@@ -192,15 +175,10 @@ func TestBootstrapWatchLogs(t *testing.T) {
 		},
 	}
 	f := gnuflag.NewFlagSet("test", gnuflag.ExitOnError)
-	f.SetOutput(s.writer)
 	command.SetFlags(f)
 	command.cloud = "aws"
 
-	ctx := &cmd.Context{
-		Context: context.Background(),
-		Stdout:  s.writer,
-	}
-
+	ctx := newTestContext(t)
 	err := command.Run(ctx)
 	c.Assert(err, qt.IsNil)
 }
@@ -218,18 +196,13 @@ func TestBootstrapFailsToGetCredential(t *testing.T) {
 		},
 	}
 	f := gnuflag.NewFlagSet("test", gnuflag.ExitOnError)
-	f.SetOutput(s.writer)
 	command.SetFlags(f)
 	command.controllerName = "controller-name"
 	command.cloud = "aws" // Need a valid cloud to reach credential error.
 	command.region = "region"
 	command.controllerVersion = "controller-version"
 
-	ctx := &cmd.Context{
-		Context: context.Background(),
-		Stdout:  s.writer,
-	}
-
+	ctx := newTestContext(t)
 	err := command.Run(ctx)
 	c.Assert(err, qt.ErrorMatches, `failed to get credential for cloud "aws": credential not found`)
 }
@@ -252,18 +225,13 @@ func TestBootstrapMultipleCredentials(t *testing.T) {
 		},
 	}
 	f := gnuflag.NewFlagSet("test", gnuflag.ExitOnError)
-	f.SetOutput(s.writer)
 	command.SetFlags(f)
 	command.controllerName = "controller-name"
 	command.cloud = "aws" // Need a valid cloud to reach credential error.
 	command.region = "region"
 	command.controllerVersion = "controller-version"
 
-	ctx := &cmd.Context{
-		Context: context.Background(),
-		Stdout:  s.writer,
-	}
-
+	ctx := newTestContext(t)
 	err := command.Run(ctx)
 	c.Assert(err, qt.ErrorMatches, `multiple credentials found for cloud "aws", please set a default or specify one using --credential`)
 
@@ -280,16 +248,6 @@ func TestBootstrapMultipleCredentials(t *testing.T) {
 		Logs:      []string{"log-line", "log-line"},
 		Watermark: 2,
 	}, nil)
-
-	s.writer.EXPECT().Write(gomock.Any()).DoAndReturn(func(b []byte) (int, error) {
-		c.Check(string(b), qt.Equals, "log-line\n")
-		return len(b), nil
-	}).Times(2)
-
-	s.writer.EXPECT().Write(gomock.Any()).DoAndReturn(func(b []byte) (int, error) {
-		c.Check(string(b), qt.Equals, "Job completed successfully.\n")
-		return len(b), nil
-	})
 
 	err = command.Run(ctx)
 	c.Assert(err, qt.IsNil)
@@ -319,7 +277,6 @@ func TestBootstrapWithDefaultCredential(t *testing.T) {
 		},
 	}
 	f := gnuflag.NewFlagSet("test", gnuflag.ExitOnError)
-	f.SetOutput(s.writer)
 	command.SetFlags(f)
 	command.controllerName = "controller-name"
 	command.cloud = cloudName
@@ -327,11 +284,7 @@ func TestBootstrapWithDefaultCredential(t *testing.T) {
 	command.controllerVersion = "controller-version"
 	command.detach = true
 
-	ctx := &cmd.Context{
-		Context: context.Background(),
-		Stdout:  s.writer,
-	}
-
+	ctx := newTestContext(t)
 	err := command.Run(ctx)
 	c.Assert(err, qt.IsNil)
 }
@@ -357,7 +310,6 @@ func TestBootstrapSpecifiedCredentialWithDefault(t *testing.T) {
 		},
 	}
 	f := gnuflag.NewFlagSet("test", gnuflag.ExitOnError)
-	f.SetOutput(s.writer)
 	command.SetFlags(f)
 	command.controllerName = "controller-name"
 	command.cloud = cloudName
@@ -366,11 +318,7 @@ func TestBootstrapSpecifiedCredentialWithDefault(t *testing.T) {
 	command.detach = true
 	command.credentialName = "cred-3" // Use a different credential than the default.
 
-	ctx := &cmd.Context{
-		Context: context.Background(),
-		Stdout:  s.writer,
-	}
-
+	ctx := newTestContext(t)
 	err := command.Run(ctx)
 	c.Assert(err, qt.ErrorMatches, `no credential found with name "cred-3"`)
 }
