+++ cmd/jaas/cmd/utils_test.go
@@ -1,15 +1,17 @@
 package cmd
 
 import (
+	"bytes"
 	"testing"
 
-	"github.com/canonical/jimm/v3/cmd/jaas/cmd/mocks"
+	jujucmd "github.com/juju/cmd/v3"
 	"go.uber.org/mock/gomock"
+
+	"github.com/canonical/jimm/v3/cmd/jaas/cmd/mocks"
 )
 
 type cmdMocks struct {
 	client *mocks.MockJIMMAPI
-	writer *mocks.MockWriter
 	store  *mocks.MockClientStore
 }
 
@@ -18,9 +20,18 @@ func setupCmdMocks(t *testing.T) *cmdMocks {
 	ctrl := gomock.NewController(t)
 	h := &cmdMocks{
 		client: mocks.NewMockJIMMAPI(ctrl),
-		writer: mocks.NewMockWriter(ctrl),
 		store:  mocks.NewMockClientStore(ctrl),
 	}
 	t.Cleanup(ctrl.Finish)
 	return h
+}
+
+func newTestContext(t *testing.T) *jujucmd.Context {
+	return &jujucmd.Context{
+		Context: t.Context(),
+		Dir:     t.TempDir(),
+		Stdin:   &bytes.Buffer{},
+		Stdout:  &bytes.Buffer{},
+		Stderr:  &bytes.Buffer{},
+	}
 }
